Login, Register, Dashboard, Onboarding tidy + middleware for existing users

+++ app/Http/Controllers/BrandsController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\Brands;
+use FixometerHelper;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
@@ -11,6 +13,10 @@ class BrandsController extends Controller
 {
 
   public function index() {
+
+    if( !FixometerHelper::hasRole(Auth::user(), 'Administrator') )
+      return redirect('/user/forbidden');
+
     $all_brands = Brands::all();
 
     return view('brands.index', [
@@ -21,6 +27,9 @@ class BrandsController extends Controller
 
   public function getCreateBrand() {
 
+    if( !FixometerHelper::hasRole(Auth::user(), 'Administrator') )
+      return redirect('/user/forbidden');
+
     return view('brands.create', [
       'title' => 'Add Brand',
     ]);
@@ -28,6 +37,9 @@ class BrandsController extends Controller
   }
 
   public function postCreateBrand(Request $request) {
+
+    if( !FixometerHelper::hasRole(Auth::user(), 'Administrator') )
+      return redirect('/user/forbidden');
 
     $name = $request->input('brand-name');
 
@@ -41,6 +53,9 @@ class BrandsController extends Controller
 
   public function getEditBrand($id) {
 
+    if( !FixometerHelper::hasRole(Auth::user(), 'Administrator') )
+      return redirect('/user/forbidden');
+
     $brand = Brands::find($id);
 
     return view('brands.edit', [
@@ -51,6 +66,9 @@ class BrandsController extends Controller
   }
 
   public function postEditBrand($id, Request $request) {
+
+    if( !FixometerHelper::hasRole(Auth::user(), 'Administrator') )
+      return redirect('/user/forbidden');
 
     $name = $request->input('brand-name');
 
@@ -63,6 +81,9 @@ class BrandsController extends Controller
   }
 
   public function getDeleteBrand($id) {
+
+    if( !FixometerHelper::hasRole(Auth::user(), 'Administrator') )
+      return redirect('/user/forbidden');
 
     Brands::find($id)->delete();
